[BE] add OpenApiConfig, method security, /me endpoint and jwt and user guard unit tests

+++ src/main/java/art/moor/ariadna/controller/UserController.java
@@ -1,5 +1,6 @@
 package art.moor.ariadna.controller;
 
+import art.moor.ariadna.config.security.UserAccessGuard;
 import art.moor.ariadna.data.dto.user.UserUpdateRequestDto;
 import art.moor.ariadna.data.dto.user.UserResponseDto;
 import art.moor.ariadna.data.model.UserRole;
@@ -7,6 +8,9 @@ import art.moor.ariadna.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,12 +22,19 @@ import java.util.UUID;
 public class UserController {
 
     private final UserService userService;
+    private final UserAccessGuard userAccessGuard;
+
+    @GetMapping("/me")
+    public UserResponseDto getMe(@AuthenticationPrincipal UserDetails userDetails) {
+        return userService.getMe(userDetails.getUsername());
+    }
 
     @GetMapping("/{id}")
     public UserResponseDto getById(@PathVariable UUID id) {
         return userService.getById(id);
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @GetMapping
     public List<UserResponseDto> getAll() {
         return userService.getAll();
@@ -32,9 +43,11 @@ public class UserController {
     @PutMapping("/{id}")
     public UserResponseDto update(@PathVariable UUID id,
                                   @Valid @RequestBody UserUpdateRequestDto request) {
+        userAccessGuard.checkCanAccess(id);
         return userService.update(id, request);
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @PutMapping("/role/{id}")
     public UserResponseDto updateRole(@PathVariable UUID id,
                                   @Valid @RequestParam UserRole role) {
@@ -43,6 +56,7 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable UUID id) {
+        userAccessGuard.checkCanAccess(id);
         userService.delete(id);
         return ResponseEntity.noContent().build();
     }
